+++ src/D20Tek.Blazor.BrowserStorage/GlobalUsings.cs
@@ -1,4 +1,5 @@
-﻿global using System.Text.Json;
+﻿global using System.Diagnostics.CodeAnalysis;
+global using System.Text.Json;
 global using Microsoft.Extensions.Options;
 global using Microsoft.JSInterop;